Move calculateAndFormatTotal and add some test coverage
